Move mitigation handling to controller level
