+++ pkg/vulnerabilityreport/controller/scanjob.go
@@ -9,10 +9,12 @@ import (
 	"github.com/aquasecurity/trivy-operator/pkg/kube"
 	"github.com/aquasecurity/trivy-operator/pkg/operator/etc"
 	. "github.com/aquasecurity/trivy-operator/pkg/operator/predicate"
+	"github.com/aquasecurity/trivy-operator/pkg/plugins/trivy"
 	"github.com/aquasecurity/trivy-operator/pkg/trivyoperator"
 	"github.com/aquasecurity/trivy-operator/pkg/vulnerabilityreport"
 	"github.com/go-logr/logr"
 	"go.uber.org/multierr"
+	"gopkg.in/yaml.v2"
 	batchv1 "k8s.io/api/batch/v1"
 	k8sapierror "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -193,6 +195,22 @@ func (r *ScanJobController) processScanJobResults(ctx context.Context, job *batc
 	if err != nil {
 		return nil, nil, err
 	}
+
+	mitigations := []Mitigation{}
+	if raw, ok := owner.GetAnnotations()[v1alpha1.VulnerabilityMitigationsAnnotation]; ok {
+		if err := yaml.UnmarshalStrict([]byte(raw), &mitigations); err != nil {
+			log.V(1).Error(err, "Failed to parse mitigations annotation", "owner", owner)
+		}
+	}
+	for i := range vulnReportData.Vulnerabilities {
+		vulnerability := vulnReportData.Vulnerabilities[i]
+		if mitigationMessage, err := GetMitigationMessage(mitigations, vulnReportData.Registry, vulnReportData.Artifact, vulnerability); err != nil {
+			log.V(1).Error(err, "Failed to check mitigations", "owner", owner)
+		} else if mitigationMessage != "" {
+			vulnReportData.Vulnerabilities[i].Mitigation = mitigationMessage
+		}
+	}
+	vulnReportData.Summary = trivy.VulnerabilitySummary(vulnReportData.Vulnerabilities)
 
 	resourceLabelsToInclude := r.GetReportResourceLabels()
